+++ packages/xrpl/HISTORY.md
@@ -10,6 +10,7 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 * Improved typescript typing
 
 ### Added
+- `getNFTokenID` lets you get the NFTokenID after minting an NFT
 
 
 ### Changed